pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "a2a-agent-selfservice"
version = "0.1.0"
description = "A2A Self-Service platform for agent creation using Google ADK, deployed on Azure"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "GeoPark Hackers" }
]
dependencies = [
    "google-adk>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "azure-identity>=1.19.0",
    "azure-keyvault-secrets>=4.8.0",
    "openai>=1.55.0",
    "structlog>=24.4.0",
    "redis>=5.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
